BAEL-7804 Replaced multiple classes with a single class with @Nested annotations to improve readability.
